refactor: Authentication
