+++ src/pages/AuthenticationPage.js
@@ -1,8 +1,10 @@
 const AuthenticationComponent = require('../components/AuthenticationComponent');
+const LoginFormComponent = require('../components/login/LoginFormComponent');
 
 class AuthenticationPage {
     constructor() {
         this.auth = AuthenticationComponent;
+        this.loginForm = LoginFormComponent;
     }
 }
