+++ src/pages/HomePage.js
@@ -1,10 +1,12 @@
 const homeComponent = require('../components/HomeComponent');
+const headerComponent = require('../components/home/HeaderComponent');
 
 
 class HomePage {
     
     constructor() {
         this.home = homeComponent;
+        this.header = headerComponent;
     }
 }
