+++ src/step-definitions/userAuthenticationSteps.js.js
@@ -1,6 +1,7 @@
 const { Given, When, Then } = require('@wdio/cucumber-framework');
 const { expect, assert } = require('chai');
 const testData = require('../utils/testData');
+const HomePage = require('../pages/HomePage');
 const AuthenticationPage = require('../pages/AuthenticationPage');
 
 Given(/^the user is on the "([^"]*)" home page$/, async (url) => {
@@ -8,14 +9,14 @@ Given(/^the user is on the "([^"]*)" home page$/, async (url) => {
 });
 
 When(/^the user clicks on "Get Trello for free"$/, async () => {
-    await AuthenticationPage.auth.clickGetTrelloForFree();
+    await HomePage.header.clickGetTrelloForFree();
 });
 
 When(/^the user enters a valid email "([^"]*)"$/, async (emailKey) => {
     const email = testData[emailKey].email;
-    await AuthenticationPage.auth.enterEmail(email);
+    await AuthenticationPage.loginForm.enterEmail(email);
 
-    const emailValue = await AuthenticationPage.auth.getEmailInputValue();
+    const emailValue = await AuthenticationPage.loginForm.getEmailInputValue();
     expect(emailValue).to.equal(email);
 });
 
@@ -28,14 +29,14 @@ When(/^the user clicks the login button$/, async () => {
 });
 
 When(/^the user clicks the continue button$/, async () => {
-    await AuthenticationPage.auth.clickContinue();
+    await AuthenticationPage.loginForm.clickContinue();
 });
 
 When(/^the user enters a valid password "([^"]*)"$/, async (emailKey) => {
     const password = testData[emailKey].password;
-    await AuthenticationPage.auth.enterPassword(password);
+    await AuthenticationPage.loginForm.enterPassword(password);
 
-    const passwordValue = await AuthenticationPage.auth.getPasswordInputValue();
+    const passwordValue = await AuthenticationPage.loginForm.getPasswordInputValue();
     expect(passwordValue).to.equal(password);
 });
